Changes x, y, width, height to double
The attribute x, y, width, height was declared as interges. They was assigned by
values where the operands was doubles. They was used in call to methods of
TexImage where the arguments expect a double value.

It must be safe to treat them as doubles all whe way.

+++ src/core/control/LatexController.h
@@ -167,22 +167,22 @@ private:
     /**
      * X-Position
      */
-    int posx = 0;
+    double posx = 0;
 
     /**
      * Y-Position
      */
-    int posy = 0;
+    double posy = 0;
 
     /**
      * Image width
      */
-    int imgwidth = 0;
+    double imgwidth = 0;
 
     /**
      * Image height
      */
-    int imgheight = 0;
+    double imgheight = 0;
 
     /**
      * Document
